backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sentence-transformers>=2.7
faiss-cpu>=1.7
pandas>=2.0
pyarrow>=14.0
pydantic>=2.7
python-dotenv>=1.0
